+++ CHANGELOG.md
@@ -7,6 +7,10 @@
 - Make Windows installer check if previous Daedalus is running ([PR 3141](https://github.com/input-output-hk/daedalus/pull/3141))
 - Fixed out-of-memory errors on https://ci.iog.io ([PR 3145](https://github.com/input-output-hk/daedalus/pull/3145))
 
+### Chores
+
+- Updated cardano-node to 8.9.1 (via cardano-wallet) ([PR 3152](https://github.com/input-output-hk/daedalus/pull/3152))
+
 ## 5.3.0
 
 ### Features